+++ src/controls/calendar.ts
@@ -1170,3 +1170,5 @@ customElements.define(tagName, WUPCalendarControl);
  *  Mar 13...14 >>>  + 1h
  *  Nov 6...7 >>> -1h
  */
+
+// todo add ability to select range